Always report provisioning job status
When DAGMan deletes the service job on its exits this job's status is
set to DELETED.  For this reason to avoid confusing users the
provisioning job status was shown by ``bps report `` only while the
workflow was running.  As a result there is no way to easily know if the
run was using automatic provisioning of the resources after it is
finished.  Made changes to always report the provisioning job status,
but if the provisioning job was deleted by DAGMan its final status is
changed to SUCCEEDED and displayed as such.